+++ corelib/src/GlobalMap.cpp
@@ -25,39 +25,63 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include <rtabmap/core/Map.h>
+#include <rtabmap/core/GlobalMap.h>
 #include <rtabmap/utilite/ULogger.h>
 #include <rtabmap/utilite/UStl.h>
 
 namespace rtabmap {
 
-Map::Map(const ParametersMap & parameters) :
+GlobalMap::GlobalMap(const ParametersMap & parameters) :
 	cellSize_(Parameters::defaultGridCellSize()),
-	fullUpdate_(Parameters::defaultGridGlobalFullUpdate()),
-	updateError_(Parameters::defaultGridGlobalUpdateError())
+	updateError_(Parameters::defaultGridGlobalUpdateError()),
+	occupancyThr_(Parameters::defaultGridGlobalOccupancyThr()),
+	probHit_(logodds(Parameters::defaultGridGlobalProbHit())),
+	probMiss_(logodds(Parameters::defaultGridGlobalProbMiss())),
+	probClampingMin_(logodds(Parameters::defaultGridGlobalProbClampingMin())),
+	probClampingMax_(logodds(Parameters::defaultGridGlobalProbClampingMax()))
 {
 	minValues_[0] = minValues_[1] = minValues_[2] = 0.0;
 	maxValues_[0] = maxValues_[1] = maxValues_[2] = 0.0;
 
-	cellSize_ = Parameters::defaultGridCellSize();
 	Parameters::parse(parameters, Parameters::kGridCellSize(), cellSize_);
 	UASSERT(cellSize_>0.0f);
 
-	Parameters::parse(parameters, Parameters::kGridGlobalFullUpdate(), fullUpdate_);
 	Parameters::parse(parameters, Parameters::kGridGlobalUpdateError(), updateError_);
 
 	UDEBUG("cellSize_           =%f", cellSize_);
-	UDEBUG("fullUpdate_         =%s", fullUpdate_?"true":"false");
 	UDEBUG("updateError_        =%f", updateError_);
+
+	// Probabilistic parameters
+	Parameters::parse(parameters, Parameters::kGridGlobalOccupancyThr(), occupancyThr_);
+	if(Parameters::parse(parameters, Parameters::kGridGlobalProbHit(), probHit_))
+	{
+		probHit_ = logodds(probHit_);
+		UASSERT_MSG(probHit_ >= 0.0f, uFormat("probHit_=%f",probHit_).c_str());
+	}
+	if(Parameters::parse(parameters, Parameters::kGridGlobalProbMiss(), probMiss_))
+	{
+		probMiss_ = logodds(probMiss_);
+		UASSERT_MSG(probMiss_ <= 0.0f, uFormat("probMiss_=%f",probMiss_).c_str());
+	}
+	if(Parameters::parse(parameters, Parameters::kGridGlobalProbClampingMin(), probClampingMin_))
+	{
+		probClampingMin_ = logodds(probClampingMin_);
+	}
+	if(Parameters::parse(parameters, Parameters::kGridGlobalProbClampingMax(), probClampingMax_))
+	{
+		probClampingMax_ = logodds(probClampingMax_);
+	}
+	UASSERT(probClampingMax_ > probClampingMin_);
 }
 
-Map::~Map()
+GlobalMap::~GlobalMap()
 {
 	clear();
 }
 
-void Map::clear(bool keepCache)
+void GlobalMap::clear(bool keepCache)
 {
+	UDEBUG("Clearing (keep cache = %d)", keepCache?1:0);
 	if(!keepCache)
 	{
 		cache_.clear();
@@ -68,7 +92,7 @@ void Map::clear(bool keepCache)
 	maxValues_[0] = maxValues_[1] = maxValues_[2] = 0.0;
 }
 
-unsigned long Map::getMemoryUsed() const
+unsigned long GlobalMap::getMemoryUsed() const
 {
 	unsigned long memoryUsage = 0;
 
@@ -91,7 +115,7 @@ unsigned long Map::getMemoryUsed() const
 	return memoryUsage;
 }
 
-void Map::addToCache(int nodeId,
+void GlobalMap::addToCache(int nodeId,
 		const cv::Mat & ground,
 		const cv::Mat & obstacles,
 		const cv::Mat & empty,
@@ -107,17 +131,13 @@ void Map::addToCache(int nodeId,
 	uInsert(cacheViewPoints_, std::make_pair(nodeId==0?-1:nodeId, viewPoint));
 }
 
-void Map::checkIfMapChanged(
-		const std::map<int, Transform> & poses,
-		bool & graphOptimized,
-		bool & graphChanged,
-		std::map<int, Transform> & transforms,
-		std::map<int, Transform> & updatedAddedNodes)
+bool GlobalMap::update(const std::map<int, Transform> & poses)
 {
-	graphOptimized = false; // If a loop closure happened (e.g., poses are modified)
-	graphChanged = addedNodes_.size()>0; // If the new map doesn't have any node from the previous map
-	transforms.clear();
-	updatedAddedNodes.clear();
+	UDEBUG("Update (poses=%d addedNodes_=%d)", (int)poses.size(), (int)addedNodes_.size());
+
+	// First, check of the graph has changed. If so, re-create the octree by moving all occupied nodes.
+	bool graphOptimized = false; // If a loop closure happened (e.g., poses are modified)
+	bool graphChanged = addedNodes_.size()>0; // If the new map doesn't have any node from the previous map
 	float updateErrorSqrd = updateError_*updateError_;
 	for(std::map<int, Transform>::iterator iter=addedNodes_.begin(); iter!=addedNodes_.end(); ++iter)
 	{
@@ -126,20 +146,70 @@ void Map::checkIfMapChanged(
 		{
 			graphChanged = false;
 			UASSERT(!iter->second.isNull() && !jter->second.isNull());
-			Transform t = Transform::getIdentity();
 			if(iter->second.getDistanceSquared(jter->second) > updateErrorSqrd)
 			{
-				t = jter->second * iter->second.inverse();
 				graphOptimized = true;
 			}
-			transforms.insert(std::make_pair(jter->first, t));
-			updatedAddedNodes.insert(std::make_pair(jter->first, jter->second));
 		}
 		else
 		{
 			UDEBUG("Updated pose for node %d is not found, some points may not be copied. Use negative ids to just update cell values without adding new ones.", jter->first);
 		}
 	}
+
+	if(graphOptimized || graphChanged)
+	{
+		// clear all but keep cache
+		clear(true);
+	}
+
+	std::list<std::pair<int, Transform> > orderedPoses;
+
+	// add old poses that were not in the current map (they were just retrieved from LTM)
+	for(std::map<int, Transform>::const_iterator iter=poses.lower_bound(1); iter!=poses.end(); ++iter)
+	{
+		if(!isNodeAssembled(iter->first))
+		{
+			UDEBUG("Pose %d not found in current added poses, it will be added to map", iter->first);
+			orderedPoses.push_back(*iter);
+		}
+	}
+
+	// insert zero after
+	if(poses.find(0) != poses.end())
+	{
+		orderedPoses.push_back(std::make_pair(-1, poses.at(0)));
+	}
+
+	if(!orderedPoses.empty())
+	{
+		assemble(orderedPoses);
+	}
+
+	//clear only negative ids
+	for(std::map<int, std::pair<std::pair<cv::Mat, cv::Mat>, cv::Mat> >::iterator iter=cache_.begin(); iter!=cache_.end();)
+	{
+		if(iter->first < 0)
+		{
+			cacheViewPoints_.erase(iter->first);
+			cache_.erase(iter++);
+		}
+		else
+		{
+			break;
+		}
+	}
+
+	return !orderedPoses.empty();
 }
+
+void GlobalMap::addAssembledNode(int id, const Transform & pose)
+{
+	if(id > 0)
+	{
+		uInsert(addedNodes_, std::make_pair(id, pose));
+	}
+}
+
 
 } // namespace rtabmap
